python/azure_openai_agent/requirements.txt
agent-framework-azure
azure-identity==1.19.0
psutil>=6.1.1
python-dotenv>=1.0.1
